feat: Refactor app state management and enhance summarization service with streaming capabilities
